+++ liblzma-rs/src/common/filter_buffer_decoder.rs
@@ -89,7 +89,7 @@ pub type lzma_code_function = Option<
 >;
 pub const __DARWIN_NULL: *mut c_void = ::core::ptr::null_mut::<c_void>();
 pub const NULL: *mut c_void = __DARWIN_NULL;
-pub const UINT64_MAX: c_ulonglong = 18446744073709551615 as c_ulonglong;
+pub const UINT64_MAX: c_ulonglong = 18446744073709551615;
 pub const LZMA_VLI_UNKNOWN: c_ulonglong = UINT64_MAX;
 #[no_mangle]
 pub unsafe extern "C" fn lzma_raw_buffer_decode(
@@ -124,7 +124,7 @@ pub unsafe extern "C" fn lzma_raw_buffer_decode(
         set_out_limit: None,
     };
     let ret_: lzma_ret = lzma_raw_decoder_init(&raw mut next, allocator, filters) as lzma_ret;
-    if ret_ as c_uint != LZMA_OK as c_uint {
+    if ret_ != LZMA_OK {
         return ret_;
     }
     let in_start: size_t = *in_pos;
@@ -140,10 +140,10 @@ pub unsafe extern "C" fn lzma_raw_buffer_decode(
         out_size,
         LZMA_FINISH,
     );
-    if ret as c_uint == LZMA_STREAM_END as c_uint {
+    if ret == LZMA_STREAM_END {
         ret = LZMA_OK;
     } else {
-        if ret as c_uint == LZMA_OK as c_uint {
+        if ret == LZMA_OK {
             if *in_pos != in_size {
                 ret = LZMA_BUF_ERROR;
             } else if *out_pos != out_size {
